many actions

+++ include/utils.h
@@ -6,9 +6,9 @@
 using namespace std;
 
 string repeat_string(string, int);
-int getStartForErrorMsg(int, string);
-int getEndForErrorMsg(int, string);
-string getContextForErrorMsg(int, string);
+int getStartForErrorMsg(size_t, string);
+int getEndForErrorMsg(size_t, string);
+string getContextForErrorMsg(size_t, string);
 string toUpperCase(string);
 string readFile(ifstream &);
 string readLine(ifstream &);
